Implement the first machine function
